pyproject.toml
[project]
name = "pigcallclassifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asgiref>=3.8.1",
    "flask>=3.1.0",
    "librosa>=0.10.2.post1",
    "matplotlib>=3.10.0",
    "numpy>=2.0.2",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "scikit-learn>=1.6.1",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "torchvision>=0.21.0",
    "uvicorn>=0.34.0",
    "pillow>=10.2.0",
]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/nightly/cpu"
explicit = true
